+++ common/src/main/java/net/mehvahdjukaar/sleep_tight/configs/ClientConfigs.java
@@ -71,7 +71,6 @@ public class ClientConfigs {
                 .define("particle_lifetime", 380, 1, 10000);
         PARTICLE_SPAWN_FREQUENCY = builder.comment("Makes particles spawn more often. Set to 0 to disable")
                 .define("particle_spawn_chance", 0.15, 0, 1);
-        TIME_FORMAT_24H = builder.define("24h_time_format", true);
         builder.pop();
 
 
@@ -82,6 +81,8 @@ public class ClientConfigs {
                 .define("crossair_insomnia_cooldown", true);
         SHOW_TIME = builder.comment("Displays current time when sleeping. WITH_CLOCK only shows it while a clock is in the inventory")
                 .define("show_time_when_sleeping", SleepTimeDisplay.ALWAYS);
+        TIME_FORMAT_24H = builder.comment("Shows the sleep time in 24h format instead of AM/PM")
+                .define("24h_time_format", true);
         VILLAGER_SLEEP = builder.comment("Makes villagers close their eyes when sleeping")
                         .define("sleeping_villagers_eyes", true);
         ZZZ_PARTICLES = builder.comment("Spawn particles when sleeping. Set to 0 to disable")